+++ app/src/main/res/layout/mars_rover_photo_activity.xml
